KUDU-3493 upgrade Guava to 32.1.1-jre
This is to address CVE-2023-2976 in 30.1-jre [1].

An update on java/build.gradle is a workaround as suggested by the
Guava release notes [2] to allow for building with gradle 6.x.

An update on build-support/verify_jars.pl allows for ProGuard [3]
rule files to be in the result JARs: those appeared in the compiled
JAR files with the new Guava version.

[1] https://nvd.nist.gov/vuln/detail/CVE-2023-2976
[2] https://github.com/google/guava/releases/tag/v32.1.0
[3] https://www.guardsquare.com/en/products/proguard

Change-Id: I4acf448085e2279be3ed8c77ccf3306494c6639c
Reviewed-on: http://gerrit.cloudera.org:8080/20235
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>
(cherry picked from commit ab2f15d0dc168245d9e5adc631784f0f1be1c803)
Reviewed-on: http://gerrit.cloudera.org:8080/20239
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>
Tested-by: Yingchun Lai <laiyingchun@apache.org>