[tests] Fix flakiness in TestAsyncKuduSession.java
The test testFlushBySize within TestAsyncKuduSession.java
has been flaky with a 39% failure rate for the last weeks.

The main reason for the failure of this case is that the
newly added unit test did not consider the impact of the
cache. I have modified the relevant code of the unit test
to verify that the newly added flush strategy is indeed
effective by performing a large number of insert operations
exceeding the number of cache buffers.

Change-Id: Iadc61fcddb9ffbfe05ef398ba61c79f8063d72de
Reviewed-on: http://gerrit.cloudera.org:8080/20238
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>