[tests] a small clean-up on create-table-itest.cc
This patch updates the code in create-table-itest.cc to fix the
compilation warning on improper handling of the return status of the
ExternalTabletServer::Restart() method.  In addition, it contains other
unsorted improvements (mostly code style-related) on the test code.

This is a follow-up to a5648f39b407c62b70ca0ce6fefd0fdab9533daf.

Change-Id: I7508f1bf0c286098c673019731c8551efd2d8d78
Reviewed-on: http://gerrit.cloudera.org:8080/20261
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Kudu Jenkins