[cmake] clean up on targets when using NO_TESTS=1
This patch updates makefiles to remove unnecessary targets when
the build is configured using 'cmake -DNO_TESTS=1 ...'.  It reduces
CPU and disk consumption when building with NO_TESTS=1.

Also, the sanitizer_options static library is now built only in
ASAN/TSAN configurations.

This patch doesn't add any functional changes in the Kudu's code itself,
except that the kudu CLI tool no longer offer the 'kudu test' action
when built with NO_TESTS=1.  That also made the size of the kudu CLI
binary a bit smaller.  If comparing the sizes of the stripped binaries,
I got the following numbers when building on CentOS7.9 with NO_TESTS=1
in RELEASE configuration using devtoolset-8:

  before: 67883424 bytes
  after:  66891680 bytes

Change-Id: I6de4f9498bef9b2102ea6507a35ad78b08d4c27a
Reviewed-on: http://gerrit.cloudera.org:8080/20281
Tested-by: Kudu Jenkins
Reviewed-by: Yifan Zhang <chinazhangyifan@163.com>
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>