[common] minor optimization on ComputeHashBuckets
It's enough to find the very first 'false' element in the boolean vector
of selected predicate values to decide whether to continue with the code
in PartitionPruner::ComputeHashBuckets().  Also, using std::all_of()
might employ optimizations in the specialized implementation of
std::vector<bool> container, if there were any.

In addition, I looked at the possibility to use boost::dynamic_bitset
instead of std::vector<bool> for PartitionPruner::PruneHashComponent()
in this context, but it seems PruneHashComponent() isn't exactly at
hot paths of reading and writing data.

This is a follow-up to 5d6774b1022f26d33abd6cc9fb9950749084942

Change-Id: Iecd9a739f408646c0dab74727f919e4f6ce3ce7e
Reviewed-on: http://gerrit.cloudera.org:8080/20288
Tested-by: Kudu Jenkins
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Reviewed-by: Yifan Zhang <chinazhangyifan@163.com>
Reviewed-by: Yuqi Du <shenxingwuying@gmail.com>
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>