KUDU-3432: Fix flakiness in client-test.cc
The 'TestGetTableKeyRange' in 'client-test.cc' may
be unstable in certain cases. Through analysis, it
was determined that the split algorithm might encounter
some corner cases when handling written data.

To mitigate the impact of these cases on CI, I have
increased the amount of data written in the unit tests
and modified some logic in dance to enhance the robustness
of the unit tests.

Change-Id: I1e3fecc09f58d91fe9a8c6d20caa21938dd9c972
Reviewed-on: http://gerrit.cloudera.org:8080/20314
Reviewed-by: Alexey Serbin <alexey@apache.org>
Tested-by: Alexey Serbin <alexey@apache.org>