[clock] clearer logging on the hybrid clock time source
This patch updates the logging messages about the time source used
for the hybrid clock's initialization.

Prior to this patch, it could be reported that the system auto-selected
a particular time source, but in fact the effective time source would be
specified explicitly via the --time_source flag and it was not 'auto',
so the time source auto-selection hadn't been involved at all.  That's
confusing, so here is a fix.

Change-Id: Ib1bdd13cd2894954002a79258a46aeaee1a44fcf
Reviewed-on: http://gerrit.cloudera.org:8080/20337
Tested-by: Kudu Jenkins
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>