[util] deprecate {MonoTime,MonoDelta}::Equals() methods
All these Equals(), LessThan(), MoreThan(), Earliest(), AddDelta(), etc.
methods look a bit funny in the C++ code.  Also, the API already exposes
more expressive means to do the same: the corresponding operators.
To prevent spreading the harder-to-read and funny looking code, this
patch declares those operators obsolete to promote using the better
option.  Of course, due to the compatibility reasons, those methods are
still kept in the public API as they used to be, but now the compiler
will emit a warning when trying to use those deprecated methods.

I also took the liberty of updating the related code to be more
style-compliant and making other minor improvements.

Change-Id: I716955dcc7359f920d0b5bedf9035a42edeb6e92
Reviewed-on: http://gerrit.cloudera.org:8080/20442
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>