KUDU-3359 Allow multi-JAR cp for Ranger client
In order to let Ranger client to write audit to HDFS, Hadoop client
JARs need to be loaded. To make sure this is possible to do, this commit
changes the behavior of the ranger_jar_path to allow colon-separated JAR
classpaths to be passed to Java.

Change-Id: If137b5541b948506372db50e984eeee7e05dfd22
Reviewed-on: http://gerrit.cloudera.org:8080/20346
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>
(cherry picked from commit 6b1240afc7e7d4c5730ff6fc61927177df607b2a)
Reviewed-on: http://gerrit.cloudera.org:8080/20416
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>