[misc] replace vector::resize() with constructor's parameter
I noticed the following pattern that seems to be replicated
in various places:

  std::vector<T> v;
  v.resize(some_constant);

This patch replaces such code with calls to one of std::vector's
constructors that accepts the size of the vectors as a parameter.

I also took the liberty of updating the related code to be more
style-compliant and making other minor improvements.

This patch doesn't contain any functional modifications.

Change-Id: Id72689597c39b66a6f2f24f734a12a90afef152a
Reviewed-on: http://gerrit.cloudera.org:8080/20484
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>