[util] clean up libunwind headers for aarch64
It turns out the specialization of libunwind headers for a particular
architecture is done within the libunwind.h header itself, so there is
no need for arch-specific preprocessor directives in the files that
use libunwind API.

I verified that the result code builds successfully on aarch64,
in particular on a Graviton3 instance in EC2.

This changelist doesn't contain any functional modifications.

This is a follow-up to f5d2899a6483fc08a17ade8d9b26b3e6930da4e4.

Change-Id: Ic2ed98f15cf83484e4f2efa8c76414d2218f12e1
Reviewed-on: http://gerrit.cloudera.org:8080/20543
Tested-by: Kudu Jenkins
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>