[thirpdarty] upgrade gperftools to 2.13
With gperftools 2.13 come many updates since 2.8.1.  At least,
as a cumulative update from 2.11rc2 (a.k.a. 2.10.80 pre-release)
comes the support for generic_fp stack collection method on aarch64
platform [1], so now there is a parity between x86_64 and aarch64 builds
in terms of available options for stack trace collection in gperftools.

Also, the libgcc-based stacktrace capturing is more robust with the new
version of gperftools.  In addition, version 2.13 includes many other
fixes since 2.8.1, for details see [2].

With the new version, by default gperftools are built with the
'-fno-omit-frame-pointer -momit-leaf-frame-pointer' flag combination
if supported by the platform, so I removed the --enable-frame-pointers
flag for the configure script.

The gperftools-tcmalloc-osx-fix.patch patch is no longer needed since
it's automatically "included" into the gperftools source code.

[1] https://github.com/gperftools/gperftools/releases/tag/gperftools-2.10.80
[2] https://github.com/gperftools/gperftools/releases

Change-Id: Ib4916bf64ea90bb148e0abf31779106a57061a15
Reviewed-on: http://gerrit.cloudera.org:8080/20545
Tested-by: Kudu Jenkins
Reviewed-by: Marton Greber <greber.mx@gmail.com>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>