KUDU-3516 No need to traverse all delta stores while computing score.
If we have many tables with many columns and each of them gets many
update requests, the maintenance scheduler thread might be stuck in
calculating the perf improvement scores of major delta compactions.
This is because we check all the updated columns while one single
updated column is actully enough.

This patch adds a new method to judge if a delta store needs to be
compacted. The scheduler will return as soon as it finds an updated
column. Please see KUDU-3516 for details.

Change-Id: I4b2c9525d8a12183130ee2cb7b5b00d1d2f60bfa
Reviewed-on: http://gerrit.cloudera.org:8080/20547
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>