[rpc] add local address into connection negotiation trace
I found it would be easier for RPC connection negotiation
troubleshooting to have the information on both the remote and the local
addresses of the connection in the negotiation trace.  That would help
to remove ambiguity when there are multiple Kudu clients running at
the same node from which RPC connections are initiated, so now it's
easier to find corresponding pairs of negotiation traces in client- and
server-side logs.

I didn't add any tests, but I manually verified that the information
on the local address is present in the connection negotiation traces,
for example:

  1011 19:59:40.775119 (+     0us) reactor.cc:618] Submitting negotiation task for client connection to 127.206.101.1:46339 (local address 127.0.0.1:47450)

  1011 19:59:40.775135 (+     0us) reactor.cc:618] Submitting negotiation task for server connection from 127.0.0.1:47450 (local address 127.206.101.1:46339)

Change-Id: I2907e9b4fd502e2139c989c940d2f1ac55d23219
Reviewed-on: http://gerrit.cloudera.org:8080/20564
Tested-by: Kudu Jenkins
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>