[Tool] Find file path where the block is located
Sometimes, a block is damaged because of the checksum is incorrect
or the disk is damaged. The log only prints the block id, but no
container id. We want to know the container where block is located.
If we know the container id, we can read the related metadata file
to get the status of the block: dead or alive. We can also truncate
the record from the metadata file to repair the container, although
it will cause some data loss.

This patch provides the tool to find the file path by a block id.

Change-Id: Ifaafd2bb634ed7fab923d9cf9eef40437442d187
Reviewed-on: http://gerrit.cloudera.org:8080/20594
Tested-by: Kudu Jenkins
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>