refactor: move SetEncryptionFlags to constructors
SetEncryptionFlags() is a test utility function used
to enable or disable encryption in unit tests. In
some parameterized tests, SetEncryptionFlags() calls
are distributed in many test cases, there is a risk
that missing to call it if add new test cases.

This patch moves the function call to the constructors,
also changes TEST_F to TEST_P for test
LogBlockManagerTest.TestContainerBlockLimitingByMetadataSizeWithCompaction

Change-Id: I67dca5a45f78de87309d0434c3b003a83f28e1a6
Reviewed-on: http://gerrit.cloudera.org:8080/20592
Tested-by: Yingchun Lai <laiyingchun@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>