KUDU-3486 Periodically send tombstoned replica report to delete it.
A tombstoned tablet replica might never be deleted since the
delete-type deletion could only occur when the tablet is deleted.
And the requests will only be sent to the voters, not including
the tombstoned ones.

The data of tombstone replica is deleted, but metadata is persisted
in memory, which will occupy a lot of memory, especially for the
tablets with big SchemaPB.

Tombstoned replicas can be deleted while processing heartbeats. So
making the tservers periodically send some reports with tombstoned
replicas will delete the useless tombstoned replicas. See KUDU-3486
for details.

Change-Id: I92d3b3e44f49ad24a95fa10f351c6bd55c6eca7b
Reviewed-on: http://gerrit.cloudera.org:8080/20600
Tested-by: Kudu Jenkins
Reviewed-by: Yifan Zhang <chinazhangyifan@163.com>