[rpc] add hostname in Messenger to use it in ServerNegotiation
This patch introduces a new 'hostname' property for the Messenger class
to use it in ServerNegotiation.  By doing so, a call to GetFQDN() is
avoided during RPC connection negotiation at the server side.  A call
to GetFQDN() might be quite expensive since sometimes it turns into
a remote call, and if DNS resolver is slow or misconfigured, the
round-trip might take several seconds.  By avoiding calls to GetFQDN()
as part of ServerNegotiation::InitSaslServer(), the server-side
negotiation is now more robust at least when using Kerberos credentials
for RPC authentication.

The hostname (usually FQDN) is retrieved as a part of KuduServer::Init()
and stored in the messenger.  It's used later on to set the 'hostname'
attribute for the server metrics and also when running server-side
RPC connection negotiation.  Of course, this is done under assumption
that the name of a node isn't changing under a running Kudu server.

This patch also contains a few test scenarios to cover the new
functionality.  Existing {TabletServerTest,MasterTest}.ServerAttributes
test scenarios have been updated accordingly.

Change-Id: I1c13e0b4c4320b5b3eeaaf8e63df8b0d56f0ee6a
Reviewed-on: http://gerrit.cloudera.org:8080/20609
Tested-by: Kudu Jenkins
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>