Fix a unit test in aarch64 system
The block size is not fixed in different architecture systems.
For example, usually on x86 architecture it is 4k but on aarch64 some
linux distribution (eg. 4.19.90-23.30.v2101.ky10.aarch6) can use a 64k
block size kernel. Test EncryptionEnabled/LogBlockManagerTest.MetricsTest/1
run fails in some aarch64 systems currently because of ignoring the
block size of the architecture. Instead of setting a fixed
value for FLAGS_log_container_max_size, it is better to use
the variable block size to set it.

Change-Id: Ide52a251b15b1af437d570c146beb0c30fed161b
Reviewed-on: http://gerrit.cloudera.org:8080/20613
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>