[fs] fix compilation on macOS
After update to the very recent code in from the main branch, compiling
the project on my macOS laptop started failing with the following error:

  Undefined symbols for architecture x86_64:
    "void google::MakeCheckOpValueString<std::nullptr_t>(std::__1::basic_ostream<char, std::__1::char_traits<char> >*, std::nullptr_t const&)", referenced from:
	std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char> >* google::MakeCheckOpString<kudu::fs::LBMCorruptor::Container const*, std::nullptr_t>(kudu::fs::LBMCorruptor::Container const* const&, std::nullptr_t const&, char const*) in log_block_manager-test-util.cc.o
  ld: symbol(s) not found for architecture x86_64

This patch addresses the issue.  I haven't dug any further w.r.t. the
difference between compiling on macOS with CLANG and compiling the same
on Linux with GCC, but I guess that's something related to the difference
in function template instantiation.

The information on the C++ compiler is below:
  $ c++ --version
  Apple clang version 13.0.0 (clang-1300.0.29.30)
  Target: x86_64-apple-darwin20.6.0
  Thread model: posix
  InstalledDir: /Applications/Xcode.13.2.1.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin

This is a follow-up to 8c5d195c340360dd962207eaa6969604c65ba9c1.

Change-Id: I36a3eb6143876a685771ea5be4227eb5cc096e75
Reviewed-on: http://gerrit.cloudera.org:8080/20618
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>