KUDU-3522 Fail early when encryption init fails
If encryption is enabled using --encrypt_data_at_rest, but the file
system was formatted without encryption enabled, Kudu would be unable to
bootstrap tablets as the encryption headers are missing, however, it
would still start happily without being able to host any tablets.

In some cases, for example when attempting to rebalance a cluster which
contain such "zombie" tablet servers, it can cause some cryptic error
message such as this:

Failed to initialize encryption: error:0607B083:digital envelope routines:EVP_CipherInit_ex:no cipher set

This commit adds a check to the FS manager to disallow starting up with
encryption enabled and existing instance files without server keys to
address this issue as it's better to fail early in this case. It also
adds a regression test to verify it works as expected.

Change-Id: I7721953869452197f537b161bd6c6b095cc7e75f
Reviewed-on: http://gerrit.cloudera.org:8080/20651
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>