[tablet] NewReplicaOpDriver() --> NewFollowerOpDriver()
This patch renames the NewReplicaOpDriver() method in the TabletReplica
class into NewFollowerOpDriver(), so it's easier to comprehend given
there is the pairing NewLeaderOpDriver() method.  That's a better name
if taking into account the terminology used throughout the project,
where a tablet has a leader and follower replicas in the sense of the
Raft protocol.

This patch doesn't contain any functional modifications.

Change-Id: I20fd90290b993ef777b017e44a36383b7f69adab
Reviewed-on: http://gerrit.cloudera.org:8080/20713
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>