[ExternalMiniCluster] improve leader_master()
ExternalMiniCluster::leader_master() function was improved to use the
RPC. However during the dns_alias_itest cases, the host that was stored
in the masters_[]->bound_rpc_hostport was a domain name, and the RPC
returns IP address and they are compared as strings. This made
leader_master() check in ExternalMiniCluster::AddTabletServer() fail.
The solution is that bound_rpc_hostport values are converted to Sockaddr
and compared to the leader_master_addr that is retrieved from the
ConnectToClusterRpc call.

ConnectToFollowerMasterTest::AuthnTokenVerifierHaveKeys was changed so
it doesn't start any tservers, since leader_master() check can't find
leader during tablet server initialization.

In master_cert_authority-itest, num of tablet servers was set to zero as
well because tablet servers are not needed for the test and waiting for
leader made the test flaky.

Change-Id: I4a40fa8e9513066d08b42eefc57a04ba90bded82
Reviewed-on: http://gerrit.cloudera.org:8080/20714
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>