[util] Add a warning into 'heap_sample_every_n_bytes' flag description.
Parameter --heap_sample_every_n_bytes is actually a parameter of
tcmalloc. Setting it to a positive value will make thread cache
request some sample allocations from page cache, which will 
compete the lock of page heap with other threads. So does
the free operation of the sample.

A tserver with high throughput might lose so much performance
because of the lock contention of tcmalloc. Almost all the apply
threads are waiting for the spin lock.

So add a warning on --heap_sample_every_n_bytes in case that
some users do not know the terrible influence of this parameter.

Change-Id: I00d55126f0cad0ac6cc130576cc45808f96f0a47
Reviewed-on: http://gerrit.cloudera.org:8080/20716
Tested-by: Kudu Jenkins
Reviewed-by: Wang Xixu <1450306854@qq.com>
Reviewed-by: Alexey Serbin <alexey@apache.org>