[tools] disambiguate parse_metrics' output
This patch improves the output of the 'kudu diagnostic parse_metrics'
CLI tool.  Prior to this patch, it was impossible to tell between
absent metric values at a particular timestamp and zeros.  With this
patch, an absent value and undefined rate of change of a metric for
a particular timestamp are indicated with the '?' ASCII character.

Change-Id: Iba0bc90a6313bd9096fa1493dd902254cdda1dc1
Reviewed-on: http://gerrit.cloudera.org:8080/20741
Tested-by: Kudu Jenkins
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>