[java] sync slow DNS resolution threshold with the C++ code
Prior to this patch, the threshold for slow DNS resolution warning
was set to just 3 ms in the code of the Kudu Java client, which didn't
make a lot of sense if talking about non-cached DNS queries.

This patch synchronizes the warning threshold with the corresponding
threshold in the Kudu's C++ code (200 ms), and also doubles the debug
threshold, bumping up the latter from 0.5 ms to 1 ms.

Change-Id: I8c2b938b2738e1127adfdf834f0e7f0176cf35b7
Reviewed-on: http://gerrit.cloudera.org:8080/20743
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>