[compaction] Fix the incorrect memory budgeting condition
Compaction budgeting code has minimum size of on-disk deltas that
is used to decide whether memory budgeting can be applied or not.
While comparing the actual on-disk size of deltas with minimum value
the fact that minimum value is in MBytes is not taken into account
and is directly compared with on-disk size which is in bytes.
The fix is to convert the min size in MBs to bytes first and then
compare.

Change-Id: I8928b15750f100785c510ee8086e5a6281b3a7b8
Reviewed-on: http://gerrit.cloudera.org:8080/20758
Tested-by: Kudu Jenkins
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Reviewed-by: Ádám Bakai <abakai@cloudera.com>
Reviewed-by: Wang Xixu <1450306854@qq.com>
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>