[rpc] extra test scenarios for exactly-once RPC logic
In the context of troubleshooting one issue, it was necessary to know
how the exactly-once RPC logic handles the re-ordering of RPC requests,
so I added these extra scenarios to make sure the implementation behaves
as expected.  I think it's worth it adding the new tests into the code
base to explicitly state how the system behaves in particular cases and
to catch possible regressions if the corresponding code modified
in the future.

Change-Id: I3eedeb786420175b604fc67a9de948cedba9c53a
Reviewed-on: http://gerrit.cloudera.org:8080/20831
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>