KUDU-3534 [compaction] Log timestamp of two matching DELETE REDO mutations.
This patch just adds an info message to log timestamp value in case
two REDO mutations of type DELETE are found to be stamped with same
time. This is an undesired condition that could possibly happen
due to corruption of mutation entries. The value will give us an
idea whether it is 0, garbled or close to some valid timestamp.

Change-Id: I508254a83046818b81db4577bf07265b46a13c9a
Reviewed-on: http://gerrit.cloudera.org:8080/20792
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Abhishek Chennaka <achennaka@cloudera.com>
Reviewed-by: Wang Xixu <1450306854@qq.com>
Reviewed-by: Alexey Serbin <alexey@apache.org>